レビュー対応, JsonParser.swiftをData下に移動
